+++ leSouk/src/main/java/com/miage/lesouk/service/AnnonceService.java
@@ -9,10 +9,10 @@ public interface AnnonceService {
     public List<Annonce> getAnnoncesCreees(Integer idUCreateur);
     public List<Annonce> getAnnoncesCandidatees(Integer idUCandidat);
     public Set<Annonce> getAnnonces(String motsCles);
-    public Annonce creerAnnonce(String nomA, String descriptionA, Long prixA,
+    public Annonce creerAnnonce(String nomA, String descriptionA, Double prixA,
             Integer idUCreateur);
     public Annonce candidaterAnnonce(Integer idA, Integer idUCandidat, 
-            Long prixCandidat);
+            Double prixCandidat);
     public Annonce cloturerAnnonce(Integer idA);
     public Annonce commenterAnnonce(Integer idA, Integer idUser, String texte);
 }
